+++ crates/client/src/client/types/gas_costs.rs
@@ -140,6 +140,9 @@ include_from_impls! {
         pub srwq: DependentCost,
         pub swwq: DependentCost,
 
+        // Non-opcode prices
+
+        pub state_root: DependentCost,
         pub contract_root: DependentCost,
     }
 }